Melhorando tratamento do calculo da idade dos users e redirec ao clicar na ft do mapa
