+++ lib/models/user.dart
@@ -35,6 +35,33 @@ class User {
       this.birthDate,
       [this.reference]);
 
+  int get yearsOld {
+    if (this.birthDate.isEmpty) {
+      return 0;
+    }
+    var birthParts = this.birthDate.split('-');
+    var birthDate = DateTime(int.parse(birthParts[0]), int.parse(birthParts[1]),
+        int.parse('${birthParts[2][0]}${birthParts[2][1]}'));
+
+    final now = new DateTime.now();
+
+    int years = now.year - birthDate.year;
+    int months = now.month - birthDate.month;
+    int days = now.day - birthDate.day;
+
+    if (months < 0 || (months == 0 && days < 0)) {
+      years--;
+      months += (days < 0 ? 11 : 12);
+    }
+
+    if (days < 0) {
+      final monthAgo = new DateTime(now.year, now.month - 1, birthDate.day);
+      days = now.difference(monthAgo).inDays + 1;
+    }
+
+    return years;
+  }
+
   Map<String, dynamic> toMap() {
     return {
       "firebaseAuthUid": firebaseAuthUid,
