+++ lib/pages/swipe_page.dart
@@ -124,30 +124,6 @@ class _SwipePage extends State<SwipePage> {
     );
   }
 
-  int _getYearOfUser(String birthDateString) {
-    var birthParts = birthDateString.split('-');
-    var birthDate = DateTime(int.parse(birthParts[0]), int.parse(birthParts[1]),
-        int.parse(birthParts[2]));
-
-    final now = new DateTime.now();
-
-    int years = now.year - birthDate.year;
-    int months = now.month - birthDate.month;
-    int days = now.day - birthDate.day;
-
-    if (months < 0 || (months == 0 && days < 0)) {
-      years--;
-      months += (days < 0 ? 11 : 12);
-    }
-
-    if (days < 0) {
-      final monthAgo = new DateTime(now.year, now.month - 1, birthDate.day);
-      days = now.difference(monthAgo).inDays + 1;
-    }
-
-    return years;
-  }
-
   Widget _makeSwipeCards() {
     return SwipeCards(
       matchEngine: _matchEngine,
@@ -194,7 +170,7 @@ class _SwipePage extends State<SwipePage> {
                                 width: 16,
                               ),
                               Text(
-                                '${_getYearOfUser(_swipeItems[index].content.birthDate)}',
+                                '${_swipeItems[index].content.yearsOld}',
                                 style: TextStyle(
                                     fontSize: 32, color: Colors.white),
                               )
